fix: remove cast to usize

+++ fs-dkr/src/add_party_message.rs
@@ -243,7 +243,7 @@ impl<E: Curve, H: Digest + Clone, const M: usize> JoinMessage<E, H, M> {
 
         #[allow(clippy::needless_range_loop)]
         for i in 0..new_n as usize {
-            for j in 1..refresh_messages.len() as usize {
+            for j in 1..refresh_messages.len() {
                 pk_vec[i] = pk_vec[i].clone()
                     + refresh_messages[j].points_committed_vec[i].clone()
                         * li_vec[j].clone();
